+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `<html-notification>`, a custom element modelled after the `Notification` API
 
+### Changed
+- Add support for `HTMLNotificationElement.vibrate` and `HTMLNotificationElement.silent`
+
 ## [v1.0.9] - 2020-07-28
 
 ### Added